requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
ccxt==4.4.12
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
discord.py==2.4.0
ffmpeg-python==0.2.0
Flask==3.0.2
fonttools==4.54.1
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
moviepy==2.1.1
mplfinance==0.12.10b0
multidict==6.0.5
mutagen==1.47.0
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pillow==10.4.0
playwright==1.43.0
priority==2.0.0
proglog==0.1.10
pycares==4.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydub==0.25.1
pyee==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyNaCl==1.5.0
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
Quandl==3.7.0
Quart==0.19.4
quart-cors==0.7.0
redis==5.2.1
requests==2.32.3
selenium==4.22.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
spotipy==2.24.0
tqdm==4.67.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.12.6
